requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
deep-translator==1.11.4
google-genai==0.6.0
google-generativeai==0.7.2